xinit: add support to lightdm
Update the .xinitrc script to support lightdm. Lightdm uses .xprofile to
start applications when the user session is started. This script is
similar to .xinitrc, but the window manager is started by the session
manager.

All the applications executed by .xinitrc were moved to .xprofile. The
.xinitrc now executes the .xprofile script to keep compatibility.

Both scripts now explicitly use bash. This avoid problems in distros
such as Ubuntu that uses dash to execute service scripts.